// src/types/barbearia.ts
export interface Agendamento extends BaseEntity {
  clienteId: string;
  funcionarioId: string;
  servicoIds: string[];
  data: Date;
  horarioInicio: string;
  horarioFim: string;
  status: StatusAgendamento;
  observacoes?: string;
  precoTotal: number;
  desconto?: number;
  formaPagamento?: string;
  pago: boolean;
}